Fix clippy warnings in arkavo-cli
- Remove unused imports (error, info) from builtin_mcp.rs
- Add #[allow(clippy::future_not_send)] for async functions using dyn Error
- Collapse nested if statements using let-chains
- Inline format string arguments
- Replace .collect::<Vec<_>>().join("") with .collect::<String>()
- Add #[allow(clippy::format_collect)] for HTML building
- Remove unnecessary u64 cast on Linux

+++ crates/arkavo-cli/src/first_run.rs
@@ -245,7 +245,7 @@ pub fn get_available_disk_space() -> f64 {
         let mut stat: MaybeUninit<libc::statvfs> = MaybeUninit::uninit();
         if libc::statvfs(path_cstr.as_ptr(), stat.as_mut_ptr()) == 0 {
             let stat = stat.assume_init();
-            let available = stat.f_bavail as u64 * stat.f_frsize;
+            let available = stat.f_bavail * stat.f_frsize;
             return available as f64 / 1_000_000_000.0;
         }
     }
